requirements.txt
django==3.2
django-environ
pytz==2021.3
sqlparse==0.4.2
psycopg2-binary==2.9.5
djangorestframework==3.14.0